<feat>: move comment-like entity to proper directory
