+++ src/comment/comment.entity.ts
@@ -1,9 +1,10 @@
 import { Entity, Column, ManyToOne, JoinColumn, OneToMany } from 'typeorm';
-import { Base } from '../common/entity/base.entity';
-import { CommentLikeEntity } from '../common/entity/comment-like.entity';
-import { CommentReplyEntity } from '../common/entity/comment-reply.entity';
+
 import { UserEntity } from '../user/user.entity';
 import { PostEntity } from '../post/post.entity';
+import { Base } from '../common/entity/base.entity';
+import { CommentLikeEntity } from '../comment-like/comment-like.entity';
+import { CommentReplyEntity } from '../common/entity/comment-reply.entity';
 
 @Entity({
   name: 'Comments',
